Split SpawnsetArena into three components and move canvas component to CanvasArena project
