+++ src/razor-core/DevilDaggersInfo.Razor.Core.CanvasArena/Components/SpawnsetArena.razor
@@ -1,0 +1,5 @@
+<div class="relative max-w-arena-table" id="arena">
+	<canvas width="@_canvasSize" height="@_canvasSize" id="arena-canvas" @ref="_canvasReference"></canvas>
+</div>
+
+<SpawnsetArenaHoverInfo @ref="_spawnsetArenaHoverInfo" />
